Rework checkin identifiers command
This commit renames the checkin identifier command to
"manageidentifiers" and now supports creating/deleting venue ids.
